+++ src/render/vulkan/compute/compute_room_texture.h
@@ -12,6 +12,6 @@ void init_compute_room_texture(const VkDevice vkDevice);
 
 void terminate_compute_room_texture(void);
 
-void computeStitchTexture(const int tilemapTextureHandle, const int destinationTextureHandle, const ImageSubresourceRange destinationRange, const Extent tileExtent, uint16_t **tileIndices);
+void computeStitchTexture(const int tilemapTextureHandle, const int destinationTextureHandle, const ImageSubresourceRange destinationRange, const Extent tileExtent, uint32_t **tileIndices);
 
 #endif	// COMPUTE_ROOM_TEXTURE_H
